Add a section for official standards; add POSIX.
Under the Language Agnostic section, I've added a heading for standards documents, and started by adding the POSIX standard, which links to some instructions I've posted on how to access it without paying. I actually have a number of these, if you think it's the type of thing that would be useful on this list?